backend/requirements.txt
django>=6.0
djangorestframework
django-cors-headers
Pillow
django-ckeditor
